implement core memory tests
