+++ core/src/context/mod.rs
@@ -1,4 +1,3 @@
-pub mod agent_memory;
 pub mod builder;
 pub mod memory;
